Refactor Kori core structure and add comprehensive TSDoc
• Move createKori implementation to _internal/core as createKoriRoot
• Extract CreateKoriOptions to separate kori-options.ts file
• Add detailed TSDoc to Kori type and createKori function
• Create RouteRegistry type for better type clarity
• Rename _internal/kori to _internal/core for clearer naming
• Simplify kori-creator.ts to thin wrapper delegating to internal implementation
