+++ packages/kori/src/_internal/core/index.ts
@@ -1,0 +1,1 @@
+export { createKoriRoot } from './kori-internal.js';
