--- packages/kori/src/_internal/kori/index.ts
@@ -1,1 +1,0 @@
-export { createKoriInternal, type KoriInternalShared } from './kori-internal.js';
