+++ packages/kori/src/kori/index.ts
@@ -1,2 +1,3 @@
 export { type Kori } from './kori.js';
-export { createKori, type CreateKoriOptions } from './kori-creator.js';
+export { createKori } from './kori-creator.js';
+export { type CreateKoriOptions } from './kori-options.js';
