+++ packages/kori/src/kori/kori.ts
@@ -7,6 +7,41 @@ import { type KoriRequestValidatorDefault } from '../request-validator/index.js'
 import { type KoriResponseValidatorDefault } from '../response-validator/index.js';
 import { type KoriRouteDefinition, type KoriRoute, type KoriRouteMethod } from '../routing/index.js';
 
+/**
+ * Kori instance providing type-safe HTTP server functionality.
+ *
+ * The Kori instance supports method chaining for configuration and provides
+ * type-safe route registration, lifecycle hooks, plugin system, and request/response
+ * validation. All type extensions are preserved through the fluent API.
+ *
+ * @template Env - Environment type containing instance-specific data
+ * @template Req - Request type with request-specific data and methods
+ * @template Res - Response type with response building capabilities
+ * @template RequestValidator - Request validator for type-safe validation
+ * @template ResponseValidator - Response validator for type-safe validation
+ *
+ * @example
+ * ```typescript
+ * const app = createKori()
+ *   .onStart(async (ctx) => {
+ *     const db = await connectDatabase();
+ *     ctx.defer(() => db.close());
+ *     return ctx.withEnv({ db });
+ *   })
+ *   .onRequest((ctx) => {
+ *     ctx.log().info('Request received', { path: ctx.req.url() });
+ *     return ctx;
+ *   })
+ *   .get('/health', (ctx) => ctx.res.text('OK'))
+ *   .post('/users', {
+ *     requestSchema: userCreateSchema,
+ *     handler: (ctx) => {
+ *       const userData = ctx.req.validatedBody();
+ *       return ctx.res.json(createUser(userData));
+ *     }
+ *   });
+ * ```
+ */
 export type Kori<
   Env extends KoriEnvironment,
   Req extends KoriRequest,
@@ -14,26 +49,28 @@ export type Kori<
   RequestValidator extends KoriRequestValidatorDefault | undefined = undefined,
   ResponseValidator extends KoriResponseValidatorDefault | undefined = undefined,
 > = {
-  // Logger
+  /** Gets the instance logger for this Kori instance */
   log(): KoriLogger;
 
-  // Lifecycle Hooks
+  /** Registers a startup hook that executes during instance initialization */
   onStart<EnvExt extends object>(
     hook: KoriOnStartHook<Env, EnvExt>,
   ): Kori<Env & EnvExt, Req, Res, RequestValidator, ResponseValidator>;
 
-  // Handler Hooks
+  /** Registers a request hook that executes before each route handler */
   onRequest<ReqExt extends object, ResExt extends object>(
     hook: KoriOnRequestHook<Env, Req, Res, ReqExt, ResExt>,
   ): Kori<Env, Req & ReqExt, Res & ResExt, RequestValidator, ResponseValidator>;
+
+  /** Registers an error hook that executes when errors occur during request processing */
   onError(hook: KoriOnErrorHook<Env, Req, Res>): Kori<Env, Req, Res, RequestValidator, ResponseValidator>;
 
-  // Plugin
+  /** Applies a plugin to extend the Kori instance with additional functionality */
   applyPlugin<EnvExt extends object, ReqExt extends object, ResExt extends object>(
     plugin: KoriPlugin<Env, Req, Res, EnvExt, ReqExt, ResExt, RequestValidator, ResponseValidator>,
   ): Kori<Env & EnvExt, Req & ReqExt, Res & ResExt, RequestValidator, ResponseValidator>;
 
-  // Child Creation
+  /** Creates a child instance with optional path prefix and configuration */
   createChild<EnvExt extends object, ReqExt extends object, ResExt extends object>(childOptions?: {
     configure: (
       kori: Kori<Env, Req, Res, RequestValidator, ResponseValidator>,
@@ -41,20 +78,27 @@ export type Kori<
     prefix?: string;
   }): Kori<Env & EnvExt, Req & ReqExt, Res & ResExt, RequestValidator, ResponseValidator>;
 
-  // Routing
+  /** Generic route registration for any HTTP method */
   route: KoriRoute<Env, Req, Res, RequestValidator, ResponseValidator>;
 
+  /** Registers a GET route */
   get: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
+  /** Registers a POST route */
   post: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
+  /** Registers a PUT route */
   put: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
+  /** Registers a DELETE route */
   delete: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
+  /** Registers a PATCH route */
   patch: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
+  /** Registers a HEAD route */
   head: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
+  /** Registers an OPTIONS route */
   options: KoriRouteMethod<Env, Req, Res, RequestValidator, ResponseValidator>;
 
-  // Fetch Handler Generation
+  /** Generates a fetch handler for deployment to runtime environments */
   generate(): KoriFetchHandler;
 
-  // Route Definitions
+  /** Returns all registered route definitions for introspection */
   routeDefinitions(): KoriRouteDefinition[];
 };
